seperated main page from registraion page and corrected name issuse
