+++ pubspec.yaml
@@ -24,6 +24,7 @@ dependencies:
   fab_circular_menu: ^0.1.5
   flutter_multiselect: ^0.4.0
   webview_flutter:
+  fluttertoast: ^4.0.0
   url_launcher: ^5.4.2
   intl: ^0.16.0 
   flutter:
